added lacking ownedByLocalPlayer checks and other fixes

+++ srcnew/Rock/Weapons/DangerWrap.cs
@@ -98,7 +98,7 @@ public class DangerWrapBubbleProj : Projectile, IDamagable {
 			if (spawnedBomb == false && ownedByLocalPlayer) {
 				Point bombPos = pos;
 
-				bomb = new Anim(bombPos, "danger_wrap_bomb", xDir, null, false);
+				bomb = new Anim(bombPos, "danger_wrap_bomb", xDir, null, true);
 				spawnedBomb = true;
 			}
 		}
@@ -148,10 +148,13 @@ public class DangerWrapBubbleProj : Projectile, IDamagable {
 	}
 
 	public override void onDestroy() {
+		base.onDestroy();
+		if (!ownedByLocalPlayer) return;
+
 		if (type == 1) bomb?.destroySelf();
 	}
 
-	public void applyDamage(float damage, Player owner, Actor actor, int? weaponIndex, int? projId) {
+	public void applyDamage(float damage, Player owner, Actor? actor, int? weaponIndex, int? projId) {
 		if (damage > 0) {
 			destroySelf();
 		}
@@ -249,7 +252,7 @@ public class DangerWrapMineProj : Projectile, IDamagable {
 			new DangerWrapExplosionProj(pos, xDir, damager.owner, damager.owner.getNextActorNetId(), true);
 		}
 	}
-	public void applyDamage(float damage, Player owner, Actor actor, int? weaponIndex, int? projId) {
+	public void applyDamage(float damage, Player owner, Actor? actor, int? weaponIndex, int? projId) {
 		health -= damage;
 		if (health <= 0) {
 			destroySelf();
@@ -419,9 +422,13 @@ public class DWrapBigBubble : Actor, IDamagable {
 		netOwner = victim;
 		this.character = victim.character ?? throw new NullReferenceException();
 		useGravity = false;
+		canBeLocal = false;
 
-		bomb = new Anim(getCenterPos(), "danger_wrap_bomb", 
-			xDir, player.getNextActorNetId(), false, true);
+		if (ownedByLocalPlayer)  {
+			bomb = new Anim(getCenterPos(), "danger_wrap_bomb", 
+				xDir, player.getNextActorNetId(), false, true);
+
+		}
 
 		netActorCreateId = NetActorCreateId.DWrapBigBubble;
 		if (rpc) {
@@ -429,7 +436,7 @@ public class DWrapBigBubble : Actor, IDamagable {
 		}
 	}
 
-	public void applyDamage(float damage, Player owner, Actor actor, int? weaponIndex, int? projId) {
+	public void applyDamage(float damage, Player owner, Actor? actor, int? weaponIndex, int? projId) {
 		health -= damage;
 		if (health <= 0) {
 			destroySelf();
@@ -453,6 +460,8 @@ public class DWrapBigBubble : Actor, IDamagable {
 
 	public override void update() {
 		base.update();
+		if (!ownedByLocalPlayer) return;
+
 		bubbleFrames++;
 
 		changePos(character.getCenterPos());
@@ -481,12 +490,15 @@ public class DWrapBigBubble : Actor, IDamagable {
 
 	public override void onDestroy() {
 		base.onDestroy();
+
+		if (!player.ownedByLocalPlayer) return;
+
 		character.bigBubble = null!;
 		if (bomb != null) bomb.destroySelf();
 		//character.removeBubble(true);
 		character.dwrapEnd();
 		character.dwrapInvulnTime = 3;
-		if (player.ownedByLocalPlayer) new Anim(pos, "danger_wrap_big_bubble_fade", xDir, player.getNextActorNetId(), true);
+		new Anim(pos, "danger_wrap_big_bubble_fade", xDir, player.getNextActorNetId(), true);
 	}
 
 	public bool isPlayableDamagable() {
